+++ docs/source/user_guide/changelog.rst
@@ -7,6 +7,7 @@ Authors since :ref:`5.0.0`: Jordi Torrents (jordi.torrentsm@gmail.com)
 - Increase and decrease GUI font size with ``Ctrl++`` and ``Ctrl+-``.
 - Fix https://gitlab.com/polavieja_lab/idtrackerai/-/issues/90
 - Optimized `"Connecting coexisting fragments"`.
+- The log file copy in session folder contains error tracebacks.
 
 5.2.11
 ======